scripts/advanced_ai_scanner.py
Nice — let’s build this end-to-end. I’ll show **which Hugging Face (HF) models / tools to use**, **why**, and give a **clear step-by-step implementation + code snippets** you can drop into your web app backend for the **Digital Footprint Risk Analyzer**. I’ll cover PII extraction, OCR, semantic matching (dark-web-like matching), classification, embeddings + vector search, integration options (local vs HF Inference API), and a simple risk-scoring recipe.

# Overview — tasks your AI must do

1. **Extract PII** (names, emails, phones, addresses, DOBs, medical terms) from free text (social posts, bios, scraped pages).
2. **OCR** text from uploaded screenshots/PDFs.
3. **Classify text** into sensitive categories (medical, financial, identity, etc.).
4. **Match user info** against leaks / a “leak corpus” (semantic / fuzzy matching).
5. **Score risk** and produce tailored remediation steps.

---

# Which HF models / libraries to use (and why)

* **PII / NER (Named Entity Recognition)**

  * **Model**: `dslim/bert-base-NER` (good, smallish, well-tested for English NER) or `dbmdz/bert-large-cased-finetuned-conll03-english` (stronger).
  * **Use**: token-level extraction of PERSON, ORG, LOC, DATE, etc. Good for structured PII extraction.

* **Zero-shot classification (detect categories without fine-tuning)**

  * **Model**: `facebook/bart-large-mnli` (standard for zero-shot).
  * **Use**: decide whether a text contains “medical condition”, “financial info”, “contact info”, etc. Useful when you don’t want to train custom classifiers.

* **Embeddings for semantic matching / fuzzy leak lookup**

  * **Model**: `sentence-transformers/all-MiniLM-L6-v2` (fast, 384-dim, great for prototyping) or `all-mpnet-base-v2` (more accurate, heavier).
  * **Use**: turn posts / leak records into vectors; use cosine similarity to find probable matches.

* **OCR (images/PDFs → text)**

  * **Option A (HF vision model):** `microsoft/trocr-base-printed` (TrOCR) for printed text.
  * **Option B (simpler / more flexible):** `pytesseract` or `easyocr` for quick local OCR (often easier for hackathon).
  * Use HF TrOCR if you want everything inside transformers; use Tesseract/EasyOCR for speed.

* **Document layout & scanned forms**

  * **Model**: `microsoft/layoutlmv3` family — good when you need structured extraction from forms / PDFs.

* **Optional summarization / explanation**

  * **Model**: `sshleifer/distilbart-cnn-12-6` or `facebook/bart-large-cnn` for short report summaries or suggested steps.

* **Serving / convenience**

  * Use **Hugging Face Inference API** (REST) to avoid running heavy models locally — great for hackathon demos. Or run the above models locally in Python if you prefer no external calls.

---

# Step-by-step implementation (hackathon-ready)

### 0) Prereqs (Python environment)

\`\`\`bash
# minimal
pip install torch transformers sentence-transformers faiss-cpu flask requests pillow pytesseract
# (if using TrOCR & vision models you may need torchvision)
\`\`\`

---

### 1) Architecture (simple)

* **Frontend (React)** → calls → **Backend (Flask/FastAPI / Node)**
* Backend orchestrates:

  1. breach-check (HaveIBeenPwned API or local sample DB)
  2. fetch user-provided public posts/handles (opt-in) or parse pasted text
  3. OCR uploaded images
  4. Run NER (PII extraction)
  5. Zero-shot classify sensitive categories
  6. Create embeddings & vector-search a leak corpus (FAISS / Supabase Vector)
  7. Compute risk score + recommendations → return report

---

### 2) PII extraction — quick Python example (HF pipeline)

\`\`\`python
from transformers import pipeline
ner = pipeline("ner", model="dslim/bert-base-NER", grouped_entities=True)

text = "My name is John Doe, email john.doe@example.com and phone +91 98 7654 3210. I live in Pune."
entities = ner(text)
# entities gives grouped spans like [{'entity_group': 'PER', 'word':'John Doe', ...}, ...]
\`\`\`

Use results to identify emails, phones, names, addresses. Combine regex checks for email/phone to complement NER.

---

### 3) Zero-shot classification (detect whether a text contains e.g. medical info)

\`\`\`python
from transformers import pipeline
classifier = pipeline("zero-shot-classification", model="facebook/bart-large-mnli")
text = "I have been diagnosed with Type 2 Diabetes and taking metformin."
labels = ["medical condition", "financial info", "contact info", "political opinion"]
out = classifier(text, candidate_labels=labels)
# Check out['labels'][0] and out['scores'][0]
\`\`\`

Combine labels + score thresholds to mark content as sensitive.

---

### 4) OCR with TrOCR (HF) or Tesseract

**TrOCR (HF)**:

\`\`\`python
from transformers import TrOCRProcessor, VisionEncoderDecoderModel
from PIL import Image

processor = TrOCRProcessor.from_pretrained('microsoft/trocr-base-printed')
model = VisionEncoderDecoderModel.from_pretrained('microsoft/trocr-base-printed')

image = Image.open("doc.png").convert("RGB")
pixel_values = processor(images=image, return_tensors="pt").pixel_values
generated_ids = model.generate(pixel_values)
text = processor.batch_decode(generated_ids, skip_special_tokens=True)[0]
\`\`\`

**Tesseract (often faster to prototype)**:

\`\`\`python
import pytesseract
from PIL import Image
text = pytesseract.image_to_string(Image.open("doc.png"))
\`\`\`

---

### 5) Embeddings + vector search (semantic “dark web” matching)

* Create an embeddings index of known leak records / test corpus (for hackathon you create a synthetic “leak DB”).
* Use `sentence-transformers` + `faiss` for nearest neighbors.

\`\`\`python
from sentence_transformers import SentenceTransformer, util
import faiss
import numpy as np

embed_model = SentenceTransformer('all-MiniLM-L6-v2')
leak_texts = ["john.doe@example.com | 9876543210 | Pune", "alice@gmail.com | ...", ...]  # sample leak records
leak_embs = embed_model.encode(leak_texts, convert_to_numpy=True)

dim = leak_embs.shape[1]
index = faiss.IndexFlatIP(dim)      # inner product (use normalized vectors for cosine)
faiss.normalize_L2(leak_embs)
index.add(leak_embs)

# Query
q = "john.doe@example.com"
q_emb = embed_model.encode([q], convert_to_numpy=True)
faiss.normalize_L2(q_emb)
D, I = index.search(q_emb, k=5)   # top 5
\`\`\`

Interpret similarity scores and treat above-threshold matches as probable leak matches.

---

### 6) Combining signals → Risk scoring (simple formula)

Pick weights and combine normalized signals:

* `breach_flag` (0/1) — if HaveIBeenPwned returns match → weight 0.4
* `pii_count` (# of unique PII types found) → weight 0.2
* `leak_similarity_score` (max cosine similarity with leak corpus, 0–1) → weight 0.25
* `public_exposure_score` (count of public posts exposing DOB/phone/etc.) → weight 0.15

Example:

\`\`\`python
risk = 0.4 * breach_flag + 0.2 * min(pii_count,5)/5 + 0.25 * leak_similarity_score + 0.15 * public_exposure_norm
risk_score = int(risk * 100)  # scale 0-100
\`\`\`

Map to Low/Medium/High: 0–39 Low, 40–69 Medium, 70+ High (tune for your demo).

---

### 7) Orchestration pseudo-flow (backend)

1. Frontend POST `/api/scan` with `{ email, name, optional handles, files[] }`.
2. Backend:

   * Consent check + rate-limiting.
   * Call breach API (HaveIBeenPwned) if email is provided.
   * If files → OCR → get text.
   * Gather any pasted text / public posts (if user supplied handles).
   * Run NER + zero-shot on all text chunks. Aggregate PII and sensitive labels.
   * Build embeddings for texts and query leak index.
   * Compute risk\_score and generate recommended steps.
3. Return JSON report to frontend for visualization.

---

### 8) Using Hugging Face Inference API (if you don't want to host models)

* Pros: no infra effort.
* Cons: rate limits / cost for many requests.

Example:

\`\`\`python
import requests
HF_TOKEN = "<YOUR_HF_TOKEN>"
API_URL = "https://api-inference.huggingface.co/models/dslim/bert-base-NER"
headers = {"Authorization": f"Bearer {HF_TOKEN}"}
resp = requests.post(API_URL, headers=headers, json={"inputs": "My email is john@example.com"})
print(resp.json())
\`\`\`

For embeddings, HF has `sentence-transformers` models as Inference API too.

---

### 9) Frontend integration (React) – brief

* User submits email/handle/files → show a progress indicator.
* Fetch `/api/scan` → show Risk Score meter, list of exposures (grouped by type), prioritized recommendations, timeline (if user scans again).

---

### 10) Privacy & legal cautions (very important)

* **Explicit consent** before scanning — get checkbox + show what you’ll search/store.
* **Do not store raw sensitive data** unless necessary. If storing, encrypt at rest (AES), and log minimal metadata.
* For third-party scans (social networks) obey their TOS — prefer user-supplied content & opt-in.
* Rate-limit, monitor abuse.
* For “dark web” matching — don’t scrape illegal sources. Use public breach APIs or synthetic dataset for demo.

---

### 11) Testing / demo data

* For hackathon, prepare a small synthetic “leak DB” (CSV) with sample leaked strings (emails + phone combos) so you can demo semantic matches without real dark web data.
* Create example social posts & scanned screenshots to demo OCR → NER flow.

---

### 12) Deployment & scaling tips

* **Hackathon:** host backend on Render/Heroku; frontend on Vercel. Use HF Inference API to avoid model hosting hassles.
* **If you want everything self-hosted later:** containerize Python microservices, use GPU if large models are used. For production vector DB use Milvus/Weaviate/Supabase Vector.

---

# Quick code bundle (minimal Flask endpoint skeleton)

\`\`\`python
from flask import Flask, request, jsonify
from transformers import pipeline
from sentence_transformers import SentenceTransformer, util

app = Flask(__name__)
ner = pipeline("ner", model="dslim/bert-base-NER", grouped_entities=True)
classifier = pipeline("zero-shot-classification", model="facebook/bart-large-mnli")
embed_model = SentenceTransformer('all-MiniLM-L6-v2')

@app.route("/api/scan", methods=["POST"])
def scan():
    data = request.json
    texts = [data.get("bio",""), data.get("posts","")]
    combined = " ".join(texts)
    ner_out = ner(combined)
    labels = classifier(combined, candidate_labels=["medical condition","financial info","contact info","identity"])
    emb = embed_model.encode([combined])
    # ... call leak index, breach API, compute risk
    report = {"ner": ner_out, "labels": labels, "risk": 42}
    return jsonify(report)
\`\`\`

---

# Final practical advice (pick these for hackathon)

* Use **Tesseract** for OCR (fast to prototype).
* Use **dslim/bert-base-NER** for PII extraction + simple regex for emails/phones.
* Use **facebook/bart-large-mnli** for zero-shot sensitive labeling.
* Use **all-MiniLM-L6-v2** embeddings + **FAISS** for demo “leak matching”.
* Use **HaveIBeenPwned** for breach checks (demo with sample data if you don’t want API signups).
* Keep everything opt-in and don’t store raw PII in logs.

---

If you want, I can:

* generate **ready-to-run Python microservice code** (full) that does OCR → NER → embeddings → FAISS matching (with synthetic leak DB), OR
* create the **React frontend mockup** to display Risk Score + recommendations, OR
* produce the **exact `requirements.txt` + sample dataset** for your hackathon demo.

Which one do you want next?
