scripts/requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
flask>=2.3.0
requests>=2.31.0
pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
supabase>=1.0.0
python-dotenv>=1.0.0
